lib/checkers: pass `debug::Settings` to checkers
