+++ src/cwe_checker_lib/src/checkers.rs
@@ -1,8 +1,10 @@
 //! The implemented CWE checks.
+//!
 //! See their module descriptions for detailed information about each check.
 //!
 //! Currently the **Memory** check is not contained in this module
-//! but directly incorporated into the [`pointer_inference`](crate::analysis::pointer_inference) module.
+//! but directly incorporated into the
+//! [`pointer_inference`](crate::analysis::pointer_inference) module.
 //! See there for detailed information about this check.
 
 /// Checkers that are supported for Linux kernel modules.
@@ -32,6 +34,58 @@ pub mod cwe_789;
 
 pub mod prelude {
     //! Prelude imports for CWE checkers.
+    pub use super::{CweModule, CweModuleFn};
+    pub use crate::utils::debug;
     pub use crate::utils::log::{CweWarning, LogMessage, WithLogs};
-    pub use crate::CweModule;
+}
+use prelude::*;
+
+use crate::pipeline::AnalysisResults;
+
+/// The generic function signature for the main function of a CWE module
+pub type CweModuleFn =
+    fn(&AnalysisResults, &serde_json::Value, &debug::Settings) -> WithLogs<Vec<CweWarning>>;
+
+/// A structure containing general information about a CWE analysis module,
+/// including the function to be called to run the analysis.
+pub struct CweModule {
+    /// The name of the CWE check.
+    pub name: &'static str,
+    /// The version number of the CWE check.
+    /// Should be incremented whenever significant changes are made to the check.
+    pub version: &'static str,
+    /// The function that executes the check and returns CWE warnings found during the check.
+    pub run: CweModuleFn,
+}
+
+impl std::fmt::Display for CweModule {
+    /// Print the module name and its version number.
+    fn fmt(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(formatter, r#""{}": "{}""#, self.name, self.version)
+    }
+}
+
+/// Get a list of all known analysis modules.
+pub fn get_modules() -> Vec<&'static CweModule> {
+    vec![
+        &crate::checkers::cwe_78::CWE_MODULE,
+        &crate::checkers::cwe_119::CWE_MODULE,
+        &crate::checkers::cwe_134::CWE_MODULE,
+        &crate::checkers::cwe_190::CWE_MODULE,
+        &crate::checkers::cwe_215::CWE_MODULE,
+        &crate::checkers::cwe_243::CWE_MODULE,
+        &crate::checkers::cwe_252::CWE_MODULE,
+        &crate::checkers::cwe_332::CWE_MODULE,
+        &crate::checkers::cwe_337::CWE_MODULE,
+        &crate::checkers::cwe_367::CWE_MODULE,
+        &crate::checkers::cwe_416::CWE_MODULE,
+        &crate::checkers::cwe_426::CWE_MODULE,
+        &crate::checkers::cwe_467::CWE_MODULE,
+        &crate::checkers::cwe_476::CWE_MODULE,
+        &crate::checkers::cwe_560::CWE_MODULE,
+        &crate::checkers::cwe_676::CWE_MODULE,
+        &crate::checkers::cwe_782::CWE_MODULE,
+        &crate::checkers::cwe_789::CWE_MODULE,
+        &crate::analysis::pointer_inference::CWE_MODULE,
+    ]
 }
