+++ src/cwe_checker_lib/src/checkers/cwe_190.rs
@@ -40,7 +40,6 @@ use crate::analysis::vsa_results::*;
 use crate::intermediate_representation::*;
 use crate::prelude::*;
 use crate::utils::symbol_utils::{get_callsites, get_symbol_map};
-use crate::CweModule;
 
 /// The module name and version
 pub static CWE_MODULE: CweModule = CweModule {
@@ -147,6 +146,7 @@ fn contains_only_non_top_absolute_value(data_domain: &DataDomain<IntervalDomain>
 pub fn check_cwe(
     analysis_results: &AnalysisResults,
     cwe_params: &serde_json::Value,
+    _debug_settings: &debug::Settings,
 ) -> WithLogs<Vec<CweWarning>> {
     let mut logs = Vec::new();
