+++ src/cwe_checker_lib/src/checkers/cwe_215.rs
@@ -23,7 +23,6 @@ use super::prelude::*;
 
 use crate::prelude::*;
 use crate::utils::log::{CweWarning, LogMessage};
-use crate::CweModule;
 
 /// The module name and version
 pub static CWE_MODULE: CweModule = CweModule {
@@ -39,6 +38,7 @@ pub static CWE_MODULE: CweModule = CweModule {
 pub fn check_cwe(
     analysis_results: &AnalysisResults,
     _cwe_params: &serde_json::Value,
+    _debug_settings: &debug::Settings,
 ) -> WithLogs<Vec<CweWarning>> {
     let binary = analysis_results.binary;
