+++ src/cwe_checker_lib/src/checkers/cwe_426.rs
@@ -76,6 +76,7 @@ fn generate_cwe_warning(sub: &Term<Sub>) -> CweWarning {
 pub fn check_cwe(
     analysis_results: &AnalysisResults,
     cwe_params: &serde_json::Value,
+    _debug_settings: &debug::Settings,
 ) -> WithLogs<Vec<CweWarning>> {
     let project = analysis_results.project;
     let config: Config = serde_json::from_value(cwe_params.clone()).unwrap();
